# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/quests": {
      "get": {
        "summary": "Get Quests",
        "description": "Get all quests, optionally filtered by status.",
        "operationId": "get_quests_quests_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quests/{quest_id}": {
      "get": {
        "summary": "Get Quest Detail",
        "description": "Get quest details.",
        "operationId": "get_quest_detail_quests__quest_id__get",
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quests/{quest_id}/team-size": {
      "patch": {
        "summary": "Update Quest Team Size",
        "description": "Update quest team size (leader only).",
        "operationId": "update_quest_team_size_quests__quest_id__team_size_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          },
          {
            "name": "team_size",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quests/{quest_id}/team-analysis": {
      "get": {
        "summary": "Get Team Analysis",
        "description": "Analyze team compatibility and skill coverage.",
        "operationId": "get_team_analysis_quests__quest_id__team_analysis_get",
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quests/{quest_id}/match/{user_id}": {
      "get": {
        "summary": "Get Quest Match Score",
        "description": "Calculate match score (Self only).",
        "operationId": "get_quest_match_score_quests__quest_id__match__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quests/{quest_id}/kick/{user_id}": {
      "post": {
        "summary": "Kick Member",
        "description": "Remove a member (leader only).",
        "operationId": "kick_member_quests__quest_id__kick__user_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quests/{quest_id}/status": {
      "post": {
        "summary": "Update Quest Status",
        "description": "Update quest status (leader only).",
        "operationId": "update_quest_status_quests__quest_id__status_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quests/{quest_id}/accept/{user_id}": {
      "post": {
        "summary": "Accept Applicant",
        "description": "Accept an applicant (leader only).",
        "operationId": "accept_applicant_quests__quest_id__accept__user_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quests/{quest_id}/complete": {
      "post": {
        "summary": "Complete Quest",
        "description": "Mark quest as completed (leader only).",
        "operationId": "complete_quest_quests__quest_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quests/{quest_id}/cancel": {
      "post": {
        "summary": "Cancel Quest",
        "description": "Cancel a quest (leader only).",
        "operationId": "cancel_quest_quests__quest_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quests/{quest_id}/start": {
      "post": {
        "summary": "Start Quest",
        "description": "Start a quest (leader only).",
        "operationId": "start_quest_quests__quest_id__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestListResponse": {
        "properties": {
          "quests": {
            "items": {
              "$ref": "#/components/schemas/QuestResponse"
            },
            "type": "array",
            "title": "Quests"
          }
        },
        "type": "object",
        "required": [
          "quests"
        ],
        "title": "QuestListResponse"
      },
...
      "QuestResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "rank": {
            "type": "string",
            "title": "Rank"
          },
          "required_skills": {
            "items": {
              "$ref": "#/components/schemas/QuestSkill"
            },
            "type": "array",
            "title": "Required Skills"
          },
          "ocean_preference": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ocean Preference"
          },
          "team_size": {
            "type": "integer",
            "title": "Team Size"
          },
          "leader_id": {
            "type": "string",
            "title": "Leader Id"
          },
          "leader_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leader Name"
          },
          "leader_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leader Class"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "applicant_count": {
            "type": "integer",
            "title": "Applicant Count",
            "default": 0
          },
          "accepted_members": {
            "items": {
              "$ref": "#/components/schemas/QuestMember"
            },
            "type": "array",
            "title": "Accepted Members",
            "default": []
          },
          "accepted_member_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accepted Member Ids",
            "default": []
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "harmony_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harmony Score",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "rank",
          "required_skills",
          "ocean_preference",
          "team_size",
          "leader_id",
          "status",
          "created_at"
        ],
        "title": "QuestResponse"
      },
...
      "UpdateStatusRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "user_id"
        ],
        "title": "UpdateStatusRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}